+++ Magic/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 # 10.8.8
 
+ - Add `catch_fish` trigger
+
  - Fix a year-old bug caused by an EffectLib change that caused a lot of effects to revert to using the
    flame particle by default
  - Fix direction bug in Velocity action when targeting an entity on the same block as the caster